ai-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-genai
python-dotenv

